chore: rename and temporarily remove hide ignored fields switch
