+++ src/components/organisms/CompareModal/InspectionActionBar.styled.tsx
